// src/helpers/fileUploader.ts
import multer from "multer";
import path from "path";
import fs from "fs";
import { v2 as cloudinary } from "cloudinary";
import { ICloudnaryResposne, IFile } from "../app/interfaces/file";

cloudinary.config({
  cloud_name: "diw2ty4qv",
  api_key: "431273688737311",
  api_secret: "0vqtGMkR20ZL2DFqT5GWWk_B2Ps",
});

const storage = multer.diskStorage({
  destination: function (req, file, cb) {
    cb(null, path.join(process.cwd(), "uploads"));
  },
  filename: function (req, file, cb) {
    cb(null, file.originalname);
  },
});

const upload = multer({ storage: storage });

const uploadToCloudinary = async (
  file: IFile
): Promise<ICloudnaryResposne | undefined> => {
  return new Promise((resolve, reject) => {
    cloudinary.uploader.upload(
      file.path,
      //  { public_id: file.originalname },
      (error: Error, result: ICloudnaryResposne) => {
        fs.unlinkSync(file.path);
        if (error) {
          reject(error);
        } else {
          resolve(result);
        }
      }
    );
  });
};

export const fileUploader = {
  upload,
  uploadToCloudinary,
};
